convert to ddk, add enterprise header with deep links, use db schema for long names

+++ db.py
@@ -1,3 +1,4 @@
+from redis import client
 import constants
 import pandas as pd
 import datetime
@@ -12,23 +13,24 @@ from cryptography.fernet import Fernet
 with open('google.key', 'rb') as file:
     secret = file.read()
 
-fernet = Fernet(os.environ.get('GOOGLE_KEY'))
+if not os.path.isfile('aw-8a5d408d-02e1-4907-9163-b4d-ed487f09f36b.json'):
+    fernet = Fernet(os.environ.get('GOOGLE_KEY'))
 
-secret_json = fernet.decrypt(secret)
+    secret_json = fernet.decrypt(secret)
 
-with open('aw-8a5d408d-02e1-4907-9163-b4d-ed487f09f36b.json', 'w') as json_file:
-    json_file.write(secret_json.decode())
-
-client = bigquery.Client()
+    with open('aw-8a5d408d-02e1-4907-9163-b4d-ed487f09f36b.json', 'w') as json_file:
+        json_file.write(secret_json.decode())
 
 def get_platforms(p_start_time, p_end_time):
+    # See get_platform_data
+    client = bigquery.Client()
     platform_query = '''
         SELECT 
             DISTINCT platform_code
         FROM
             `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`
         WHERE
-            time>='{}' and time<='{}'
+            observation_date>='{}' and observation_date<='{}'
         ORDER BY
             platform_code
     '''.format(p_start_time, p_end_time)
@@ -39,16 +41,23 @@ def get_platforms(p_start_time, p_end_time):
         print(e)
         return None
 
-def get_platform_data(start_time, end_time, in_platform_code):
+def get_platform_data(in_platform_code):
+
+    # I used to reuse the client, but in certain situations, the second call failed with:
+    # 404 GET https://bigquery.googleapis.com/bigquery/v2/projects/aw-8a5d408d-02e1-4907-9163-b4d/jobs/5f872090-dd9d-40d6-96b8-b60c05fb05c3?projection=full&prettyPrint=false: Not found: Job aw-8a5d408d-02e1-4907-9163-b4d:5f872090-dd9d-40d6-96b8-b60c05fb05c3
+    #
+    # Rebuilding the client before every call fixed that issue.
+    
+    client = bigquery.Client()
+
     platform_constraint = '("' + '","'.join(in_platform_code) + '")'
-    platform_data = '''
+    platform_data = f'''
         SELECT
             *
         FROM
             `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`
-        WHERE platform_code IN {}
-    '''.format(platform_constraint)
-
+        WHERE platform_code IN {platform_constraint}
+    '''
     try:
         df = client.query(platform_data).to_dataframe()
         df['platform_code'] = df['platform_code'].astype(str)
@@ -59,6 +68,8 @@ def get_platform_data(start_time, end_time, in_platform_code):
      
         
 def get_summary_for_platform(start_time, end_time, in_platform_code):
+    # See comment in get_platform_data
+    client = bigquery.Client()
     platform_constraint = '("' + '","'.join(in_platform_code) + '")'
     platform_summary = '''
         SELECT
@@ -73,7 +84,7 @@ def get_summary_for_platform(start_time, end_time, in_platform_code):
         WHERE ST_CONTAINS(
             grid_cells.geometry,
             ST_GeogPoint(obs.longitude, obs.latitude)
-        ) AND obs.time>='{}' AND obs.time<='{}' AND platform_code IN {}
+        ) AND obs.observation_date>='{}' AND obs.observation_date<='{}' AND platform_code IN {}
         GROUP BY gid
         ORDER BY gid
     '''.format(start_time, end_time, platform_constraint)
@@ -86,7 +97,8 @@ def get_summary_for_platform(start_time, end_time, in_platform_code):
 
 
 def counts_by_week(start_time, end_time, parameter):
-
+    # See comment in get_platform_data
+    client = bigquery.Client()
     week_count = '''  
             SELECT
                 geo_id AS gid,
@@ -95,14 +107,14 @@ def counts_by_week(start_time, end_time, parameter):
                 ANY_VALUE(geometry) as cell,
                 platform_type,
                 COUNT('{}') AS obs,
-                EXTRACT(WEEK from time) as week
+                EXTRACT(WEEK from observation_date) as week
             FROM
                 `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations` AS nobs,
                 `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.grid-5-by-5` AS grid_cells
             WHERE ST_CONTAINS(
                 grid_cells.geometry,
                 ST_GeogPoint(nobs.longitude, nobs.latitude)
-            ) and nobs.time>='{}' and nobs.time<='{}' and EXTRACT(week from time) > 0 AND nobs.{} IS NOT NULL
+            ) and nobs.observation_date>='{}' and nobs.observation_date<='{}' and EXTRACT(week from observation_date) > 0 AND nobs.{} IS NOT NULL
             GROUP BY gid, platform_type, week
             ORDER BY gid     
         '''.format(parameter, start_time, end_time, parameter)
@@ -115,6 +127,8 @@ def counts_by_week(start_time, end_time, parameter):
 
 
 def get_summary(start_time, end_time):
+    # See comment in get_platform_data
+    client = bigquery.Client()
     summary = '''
         SELECT
             geo_id AS gid,
@@ -129,7 +143,7 @@ def get_summary(start_time, end_time):
         WHERE ST_CONTAINS(
             grid_cells.geometry,
             ST_GeogPoint(obs.longitude, obs.latitude)
-        ) and obs.time>='{}' and obs.time<='{}'
+        ) and obs.observation_date>='{}' and obs.observation_date<='{}'
         GROUP BY gid, platform_type
         ORDER BY gid
     '''.format(start_time, end_time)
@@ -142,14 +156,16 @@ def get_summary(start_time, end_time):
         return None, None
 
 
-def get_data_from_bq(platform, time0, time1): 
+def get_data_from_bq(platform, time0, time1):
+    # See get_platform_data
+    client = bigquery.Client()
     try:
         t0 = time.time()
-        sql = 'SELECT * FROM `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations` WHERE platform_code="' + platform + '" AND time>="' + time0 + '" AND time<"' + time1 +'" ORDER BY `time`'
+        sql = 'SELECT * FROM `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations` WHERE platform_code="' + platform + '" AND observation_date>="' + time0 + '" AND observastion_date<"' + time1 +'" ORDER BY `observation_date`'
         df = client.query(sql).to_dataframe()
         t1 = time.time()
-        df.loc[:,'millis'] = pd.to_datetime(df['time']).view(np.int64)
-        df.loc[:,'text_time'] = df['time'].astype(str)
+        df.loc[:,'millis'] = pd.to_datetime(df['observation_date']).view(np.int64)
+        df.loc[:,'text_time'] = df['observation_date'].astype(str)
         df.loc[:,'trace_text'] = df['text_time'] + "<br>" + df['platform_type'] + "<br>" + df['country'] + "<br>" + df['platform_code']
         t2 = time.time()
         # print ("Read to DataFrame: %.4f | Add columns: %.4f | Rows %d " % (t1-t0, t2-t1, df.shape[0]))
@@ -159,8 +175,10 @@ def get_data_from_bq(platform, time0, time1):
 
 
 def get_platform_info(platform):
+    # See get_platform_data
+    client = bigquery.Client()
     sql = '''SELECT * FROM `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.metadata` WHERE WMO="{}" 
-             and TIME=(SELECT max(time) from `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.metadata` where WMO="{}")
+             and observation_date=(SELECT max(observation_date) from `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.metadata` where WMO="{}")
           '''
     sql = sql.format(platform, platform)
     df = client.query(sql).to_dataframe()
@@ -169,7 +187,7 @@ def get_platform_info(platform):
         return df
     columns = list(df.columns)
     columns.remove('WMO')
-    columns.remove('time')
+    columns.remove('observation_date')
     df_combo = df.groupby('WMO')[columns].agg(lambda col: '; '.join(col.unique())).reset_index()
     df_cols = pd.DataFrame(columns=['Parameter', 'Value'])
     df_cols['Parameter'] = list(df_combo.columns)
@@ -178,21 +196,23 @@ def get_platform_info(platform):
 
 
 def get_time_range_locations_from_bq(time0, time1):
+    # See get_platform_data
+    client = bigquery.Client()
     sql="""
     SELECT * FROM 
     `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`
     INNER JOIN
-    (SELECT DISTINCT platform_code, max(`time`) as maxt, min(observation_depth) as minod FROM `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`  WHERE time>"{}" AND time<"{}" GROUP BY platform_code) as findme
+    (SELECT DISTINCT platform_code, max(`observation_date`) as maxt, min(observation_depth) as minod FROM `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`  WHERE observation_date>"{}" AND time<"{}" GROUP BY platform_code) as findme
     on
     `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`.platform_code=findme.platform_code and
-    `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`.time=findme.maxt and
+    `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`.observation_date=findme.maxt and
     `aw-8a5d408d-02e1-4907-9163-b4d.OSMC.observations`.observation_depth=findme.minod
     """
     tr_q = sql.format(time0, time1)
     df = client.query(tr_q).to_dataframe()
     df = df.dropna(subset=['latitude','longitude'], how='any')
     df.query('-90.0 <= latitude <= 90', inplace=True)
-    df.loc[:,'millis'] = pd.to_datetime(df['time']).view(np.int64)
-    df.loc[:,'text_time'] = df['time'].astype(str)
+    df.loc[:,'millis'] = pd.to_datetime(df['observation_date']).view(np.int64)
+    df.loc[:,'text_time'] = df['observation_date'].astype(str)
     df.loc[:,'trace_text'] = df['text_time'] + "<br>" + df['platform_type'] + "<br>" + df['country'] + "<br>" + df['platform_code']
     return df
